Make Hornix a standalone executable (only depending on LLVM)
Instead of a FunctionPass inside LLVM, we need more control over what
the translation is doing, and how it is interacting with the solver.
Thre previous architecture did not allow for information flow in the
reverse direction, which we need for producing witnesses.

The new architecture makes hornix a standalone tool, depending on LLVM
library.
It takes as input an LLVM IR file (.ll). This should be easily extended
to support also bitcode files (.bc). In the future, we should figure
out how to take as input the original C file as well.

When the IR is parsed, we get out an LLVM Module. The translation to CHC
operates on Module, and computes the SMT-LIB text representation of the
CHC system.
This is then passed to a backend Horn solver. The solver is run as a
separate process. Currently, the communication happens using temporary
files. This part is very rough and requires more work.

At the end, we get back a response from the solver, which we currently
just print to output, but we can translate it to desired format in the
future.