Improve encoding to use smaller predicates
The predicates should encode only the information necessary, they do not
need to carry along local variables through the whole function.

We more carefully compute information about which local variable needs
to propagate across basic blocks (liveness) and eliminate the need for
prime variables (this is possible because we are encoding SSA
representation).